Complete exercises 1 and 2
Remember, to allow the header's background color
to partially show, you need to use a margin. Padding
would have provided a similar effect, but it would
have applied to the left-links and right-links divs.

Also remember, that for a flex container to expand
when it is already a child element itself, its
parent needs to be a flex container too.