Complete exercise 4
Remember, when implementing a design like an
image above text, "flex-direction: column" is not
necessary (assuming that the container the 2
elements exist in is not a flex container).
Likely due to the image being block display-type.

Also, note that "text-align: center" also applies
and works on image flex items, not just text
flex items.

Finally, note that the "max-width" property will
cause text to wrap once that limit is reached.