Complete exercise 6
Note how the exercise used "<meta name="viewport"
content="width=device-width, initial-scale=1.0"
/>". It allows for the website's width to adjust
to the screen-width of the user's device (does
this make text wrap automatically onto the next
line when it overflows horizontally?).

Also note the use of "height: 100vh;" for the body
in the CSS. It means the height of the body is
equal to 100% of the viewport height (user device
screen height). This helped me when I set the
".content" class of the page to "flex-direction:
column". The ".content" class content was able to
expand its height to the full page (as was needed)
when set to "flex: 1 1 0"